server/requirements.txt
# Save as: server/requirements.txt

Flask==2.0.1
Flask-CORS==3.0.10
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.1
numpy==1.21.1
gunicorn==20.1.0